# openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/authn/exchange:
        post:
            tags:
                - AuthnService
            description: "Exchange exchanges an OAuth authorization code for IdP tokens.\r\n\r\n Supports PKCE via the optional `code_verifier` field for public clients\r\n (SPAs, mobile apps) that cannot safely store a client secret. When the\r\n authorization request was made with a `code_challenge`, the same\r\n `code_verifier` MUST be supplied here.\r\n\r\n NOTE: kernel authn exposes CodeVerifier on AuthCodeExchangeRequest, but\r\n the current Casdoor SDK adapter (kernel/authn/casdoor/token.go) does not\r\n yet forward it to GetOAuthToken. The field is accepted on the wire for\r\n forward compatibility; verify the kernel version before relying on PKCE\r\n enforcement server-side."
            operationId: AuthnService_Exchange
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/aisphere.hub.authn.v1.ExchangeRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/aisphere.hub.authn.v1.TokenResponse'
    /v1/authn/introspect:
        post:
            tags:
                - AuthnService
            description: "Introspect validates a token and returns the principal it represents.\r\n\r\n Useful for debug tools, third-party integrations, and service-to-service\r\n calls that need to verify a token's validity and identity without going\r\n through the full request middleware chain. Mirrors RFC 7662 token\r\n introspection semantics (active=false when the token is invalid,\r\n expired, or revoked)."
            operationId: AuthnService_Introspect
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/aisphere.hub.authn.v1.IntrospectRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/aisphere.hub.authn.v1.IntrospectResponse'
    /v1/authn/login-url:
        get:
            tags:
                - AuthnService
            description: "LoginURL returns the IdP authorization URL as JSON for SPA/debug usage.\r\n\r\n Browser users should prefer GET /v1/authn/login, which is registered as a\r\n direct HTTP 302 route by the service implementation. See the service\r\n comment above."
            operationId: AuthnService_LoginURL
            parameters:
                - name: redirectUri
                  in: query
                  schema:
                    type: string
                - name: state
                  in: query
                  schema:
                    type: string
                - name: scope
                  in: query
                  schema:
                    type: string
                - name: prompt
                  in: query
                  schema:
                    type: string
                - name: orgId
                  in: query
                  description: "org_id optionally overrides the globally configured Casdoor organization\r\n for this login URL. When unset, the global config organization is used.\r\n Maps to kernel authn.LoginURLRequest.OrgID."
                  schema:
                    type: string
                - name: appId
                  in: query
                  description: "app_id optionally overrides the globally configured Casdoor application.\r\n Maps to kernel authn.LoginURLRequest.AppID."
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/aisphere.hub.authn.v1.LoginURLResponse'
    /v1/authn/logout-url:
        get:
            tags:
                - AuthnService
            description: "LogoutURL returns the IdP logout URL as JSON for SPA/debug usage.\r\n\r\n Browser users can also call GET /v1/authn/logout for a direct HTTP 302.\r\n\r\n Side effect: token revocation. When the request carries an\r\n `Authorization: Bearer <access_token>` header, OR an explicit\r\n `access_token` field is supplied in the request body, the implementation\r\n calls Revoke on that access token before returning the IdP logout URL.\r\n This adds the token to the local session blacklist so subsequent API\r\n calls with the same token are rejected even before the IdP session\r\n expires. The `access_token` field makes this previously-implicit side\r\n effect visible in the API contract.\r\n\r\n NOTE: kernel authn does not currently expose a LogoutURLBuilder\r\n interface. The hub data layer must construct the Casdoor logout URL\r\n itself using the configured Casdoor endpoint, client_id and\r\n post_logout_redirect_uri. See legacy backend internal/data/auth.go for\r\n the pattern (it built the URL via the old aisphere-kit authn adapter;\r\n the new hub will need to call casdoor SDK directly or compose the URL\r\n from config)."
            operationId: AuthnService_LogoutURL
            parameters:
                - name: postLogoutRedirectUri
                  in: query
                  schema:
                    type: string
                - name: idTokenHint
                  in: query
                  schema:
                    type: string
                - name: state
                  in: query
                  schema:
                    type: string
                - name: accessToken
                  in: query
                  description: "access_token explicitly identifies the token to revoke locally before\r\n returning the IdP logout URL. When unset, the implementation falls back\r\n to the `Authorization: Bearer <access_token>` request header.\r\n\r\n This field makes the implicit revocation side effect of LogoutURL\r\n visible in the API contract instead of being hidden in service code\r\n (legacy backend internal/biz/auth.go:LogoutURL).\r\n\r\n Revoke is best-effort. If the local session manager is unavailable the\r\n implementation MUST still return the logout URL and log a warning."
                  schema:
                    type: string
                - name: orgId
                  in: query
                  schema:
                    type: string
                - name: appId
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/aisphere.hub.authn.v1.LogoutURLResponse'
...
    /v1/authn/refresh:
        post:
            tags:
                - AuthnService
            description: Refresh refreshes an IdP access token using a refresh token.
            operationId: AuthnService_Refresh
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/aisphere.hub.authn.v1.RefreshRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/aisphere.hub.authn.v1.TokenResponse'
    /v1/authn/revoke:
        post:
            tags:
                - AuthnService
            description: "Revoke explicitly revokes an access or refresh token.\r\n\r\n Use this for multi-device session management (e.g. \"sign out everywhere\"\r\n flows) where the caller wants to invalidate a token without navigating\r\n the IdP logout page.\r\n\r\n IdP-side revocation is best-effort; the local session blacklist is\r\n always updated so the token is rejected at the next API call regardless\r\n of whether the IdP actually revokes it.\r\n\r\n NOTE: the current Casdoor SDK adapter (kernel/authn/casdoor/token.go)\r\n returns ErrIdentityBackendFailed(\"casdoor sdk does not expose token\r\n revocation\"). Until the kernel adapter is updated, this RPC will only\r\n update the local blacklist and return a warning. Callers that need\r\n full IdP-side revocation should use LogoutURL with id_token_hint."
            operationId: AuthnService_Revoke
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/aisphere.hub.authn.v1.RevokeRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/aisphere.hub.authn.v1.RevokeResponse'
...
components:
    schemas:
...
        aisphere.hub.authn.v1.ExchangeRequest:
            required:
                - code
                - redirectUri
            type: object
            properties:
                code:
                    type: string
                redirectUri:
                    type: string
                state:
                    type: string
                codeVerifier:
                    type: string
                    description: "code_verifier is the PKCE code verifier. Required when the authorization\r\n request was made with a code_challenge; recommended for all public\r\n clients (SPAs, mobile apps) that cannot store a client secret.\r\n Maps to kernel authn.AuthCodeExchangeRequest.CodeVerifier.\r\n\r\n NOTE: see Exchange RPC comment — the current Casdoor adapter does not\r\n yet forward this to the SDK. The field is accepted on the wire for\r\n forward compatibility."
                orgId:
                    type: string
                appId:
                    type: string
        aisphere.hub.authn.v1.IntrospectRequest:
            required:
                - token
            type: object
            properties:
                token:
                    type: string
                    description: token is the token to validate and inspect.
                tokenType:
                    type: string
                    description: "token_type hints which kind of token is being introspected. Accepted\r\n values: \"access_token\" (default), \"refresh_token\", \"id_token\".\r\n Maps to kernel authn.VerifyTokenRequest.TokenType."
                issuer:
                    type: string
                    description: "issuer optionally restricts the accepted token issuer. When unset, the\r\n globally configured IdP issuer is used."
                audience:
                    type: array
                    items:
                        type: string
                    description: "audience optionally restricts the accepted token audience. When unset,\r\n the audience check is skipped."
                orgId:
                    type: string
                appId:
                    type: string
        aisphere.hub.authn.v1.IntrospectResponse:
            type: object
            properties:
                active:
                    type: boolean
                    description: "active indicates whether the token is currently valid (not expired, not\r\n revoked, signature OK, issuer/audience match if specified)."
                principal:
                    $ref: '#/components/schemas/aisphere.hub.authn.v1.Principal'
                scope:
                    type: string
                    description: scope is the space-delimited scope list attached to the token.
                expiresAt:
                    type: string
                    description: expires_at is the token's expiration time. Zero value means unknown.
                    format: date-time
                issuedAt:
                    type: string
                    description: issued_at is the token's issue time. Zero value means unknown.
                    format: date-time
            description: IntrospectResponse mirrors RFC 7662 token introspection.
        aisphere.hub.authn.v1.LoginURLResponse:
            type: object
            properties:
                loginUrl:
                    type: string
        aisphere.hub.authn.v1.LogoutURLResponse:
            type: object
            properties:
                logoutUrl:
                    type: string
...
        aisphere.hub.authn.v1.Principal:
            type: object
            properties:
                subjectType:
                    type: string
                subjectId:
                    type: string
                username:
                    type: string
                name:
                    type: string
                displayName:
                    type: string
                email:
                    type: string
                avatar:
                    type: string
                orgId:
                    type: string
                projectId:
                    type: string
                roles:
                    type: array
                    items:
                        type: string
                groups:
                    type: array
                    items:
                        type: string
                isAuthenticated:
                    type: boolean
                phone:
                    type: string
                    description: "phone is the user's phone number when available from the IdP.\r\n Maps to kernel authn.Principal.Phone."
                scopes:
                    type: array
                    items:
                        type: string
                    description: "scopes are the OAuth/OIDC scopes attached to the current access token.\r\n Frontends can use this to gate UI affordances (e.g. hide \"write\"\r\n buttons when the token lacks the write scope).\r\n Maps to kernel authn.Principal.Scopes."
                authMethod:
                    type: string
                    description: "auth_method records how the subject authenticated. Common values match\r\n the kernel authn.AuthMethod* constants: \"oauth\", \"oidc\", \"jwt\",\r\n \"api_key\", \"mtls\", \"session\", \"dev_token\", \"none\".\r\n Maps to kernel authn.Principal.AuthMethod."
                expiresAt:
                    type: string
                    description: "expires_at is when the current access token expires. Frontends should\r\n use this to schedule proactive token refreshes before the token\r\n expires, rather than waiting for a 401 to trigger a reactive refresh.\r\n Zero value means the expiry is unknown (e.g. the IdP did not return\r\n an exp claim).\r\n Maps to kernel authn.Principal.ExpiresAt."
                    format: date-time
                attributes:
                    type: object
                    description: "attributes carries provider-specific claims (e.g. Casdoor owner/name,\r\n external ID, signin method, custom fields) that are not part of the\r\n canonical Principal schema. Values must be JSON-serializable.\r\n Maps to kernel authn.Principal.Attributes.\r\n\r\n NOTE: some attributes (such as `access_token`) are sensitive and\r\n SHOULD be filtered out by the service layer before returning to\r\n frontend clients. See kernel/authn/casdoor/token.go:principalFromClaims\r\n for the list of attributes currently set by the Casdoor adapter."
            description: "Principal is the normalized identity of the current caller.\r\n\r\n Field set is aligned with kernel authn.Principal (github.com/aisphereio/\r\n kernel/authn.Principal). The kernel type has additional internal fields\r\n (Provider, ExternalID, Issuer, Audience, TenantID, AppID, IssuedAt) that\r\n are not exposed here because they are not useful to typical frontend\r\n consumers; if a future use case needs them, add fields rather than\r\n exposing the kernel struct verbatim."
        aisphere.hub.authn.v1.RefreshRequest:
            required:
                - refreshToken
            type: object
            properties:
                refreshToken:
                    type: string
                scope:
                    type: string
                orgId:
                    type: string
                appId:
                    type: string
        aisphere.hub.authn.v1.RevokeRequest:
            required:
                - token
            type: object
            properties:
                token:
                    type: string
                    description: token is the access, refresh, or id token to revoke.
                tokenType:
                    type: string
                    description: "token_type hints which kind of token is being revoked. Accepted values:\r\n   \"access_token\"  (default)\r\n   \"refresh_token\"\r\n   \"id_token\"\r\n Maps to kernel authn.RevokeTokenRequest.TokenType."
                orgId:
                    type: string
                appId:
                    type: string
            description: RevokeRequest asks the implementation to revoke a single token.
        aisphere.hub.authn.v1.RevokeResponse:
            type: object
            properties:
                revoked:
                    type: boolean
                    description: "revoked is true when the token was successfully revoked (either IdP-side\r\n or local blacklist)."
                idpRevoked:
                    type: boolean
                    description: "idp_revoked is true when the IdP itself confirmed revocation. False\r\n means only the local session blacklist was updated; the token may still\r\n be valid against other Resource Servers until it expires."
                message:
                    type: string
                    description: "message carries a human-readable note, e.g. \"casdoor sdk does not expose\r\n token revocation; only local blacklist updated\"."
            description: "RevokeResponse returns the outcome of a Revoke call.\r\n\r\n We do not use google.protobuf.Empty so callers can distinguish\r\n \"IdP-side revoked\" from \"only local blacklist updated\" — important for\r\n security-sensitive flows where the caller needs to know whether the token\r\n may still be valid against other Resource Servers."
        aisphere.hub.authn.v1.TokenResponse:
            type: object
            properties:
                accessToken:
                    type: string
                refreshToken:
                    type: string
                idToken:
                    type: string
                tokenType:
                    type: string
                expiresIn:
                    type: string
                scope:
                    type: string
            description: "TokenResponse is the OAuth2/OIDC token set returned by Exchange and Refresh.\r\n\r\n `expires_in` is seconds from issue time to expiry (OAuth2 standard). For a\r\n wall-clock expiry, callers can compute issued_at + expires_in, or use the\r\n `expires_at` field on Principal returned by Me/Introspect."
...